+++ R/csvFileInput.R
@@ -132,13 +132,14 @@ csvFile <- function(input, output, session) {
     #  names(out) <- gsub(x[1], x[2], names(out))
     #}
 
+
     numstart.vnum <- suppressWarnings(sapply(names(out),function(x){!is.na(as.numeric(substr(x, 1,1)))}))
     names(out)[numstart.vnum] <- paste("n_", names(out)[numstart.vnum], sep = "")
 
     factor_vars <- names(out)[out[, lapply(.SD, class) %in% c("factor", "character")]]
     out[, (factor_vars) := lapply(.SD, as.factor), .SDcols= factor_vars]
     conti_vars <- setdiff(names(out), factor_vars)
-    nclass <- unlist(out[, lapply(.SD, function(x){length(unique(x))}), .SDcols = conti_vars])
+    nclass <- unlist(out[, lapply(.SD, function(x){length(unique(x)[!is.na(unique(x))])}), .SDcols = conti_vars])
     #except_vars <- names(nclass)[ nclass== 1 | nclass >= 10]
     add_vars <- names(nclass)[nclass >= 1 &  nclass <= 5]
     #factor_vars_ini <- union(factor_vars, add_vars)